+++ packages/ergo-wasm-common/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2021"
 [dependencies]
 wasm-bindgen = "0.2.63"
 js-sys = "0.3.60"
+with_builtin_macros = "0.0.3"
