+++ packages/ergo-wasm-common/src/decls.rs
